style: move to the set utility
